refactor: refactor parse code
